refactor(fund): fold Systems.Budget into Systems.Fund
Systems.Budget owned pay-in mechanics (transactions, currency ledger,
expiration, reconciliation); Systems.Fund owned rewards + payouts.
Conceptually one money umbrella, two namespaces. Fold Budget into Fund
per Epic OPP Phase 2's declared architecture ("Systems.Fund replaces
Systems.Budget").

- Merged Systems.Budget.Public into Systems.Fund.Public (pay-in creation,
  transaction completion / failure / expiration, reconciliation
  delegations).
- Moved Budget modules to Fund at same basenames: TransactionModel,
  CurrencyLedgerModel, PayInRequestModel, PayInRequestForm,
  PayInExpirationWorker, TransactionReconciliation,
  TransactionOrphanReconciliation. (git detects as renames.)
- Renamed Systems.Budget.* → Systems.Fund.* across all callers (payment
  controllers + reconciliation workers, assignment paid_slots_logic +
  _assembly, fund model, seeds, factories, cron config in dev.exs +
  runtime.fly.exs).
- Tests followed to test/systems/fund/.

Table names (`transactions`, `currency_ledger`) unchanged — this is a
namespace-only fold; no DB migration. The rename to `fund_*` table names
is deferred (out of scope, low value on its own).

Fixes https://app.basecamp.com/5734045/buckets/35926565/todos/10169488204
