+++ core/systems/fund/transaction_reconciliation.ex
@@ -1,4 +1,4 @@
-defmodule Systems.Budget.TransactionReconciliation do
+defmodule Systems.Fund.TransactionReconciliation do
   @moduledoc """
   Reconciliation for pay-in transactions. Re-applies the provider's current
   status to `:pending`/`:failed` transactions whose webhook was lost (including
@@ -14,7 +14,7 @@ defmodule Systems.Budget.TransactionReconciliation do
   require Logger
 
   alias Core.Repo
-  alias Systems.Budget
+  alias Systems.Fund
   alias Systems.Payment
   alias Systems.Payment.ReconciliationState, as: State
 
@@ -38,7 +38,7 @@ defmodule Systems.Budget.TransactionReconciliation do
     age_cutoff = NaiveDateTime.add(now, -min_age_minutes * 60, :second)
     lookback_cutoff = NaiveDateTime.add(now, -max_age_days * 24 * 60 * 60, :second)
 
-    from(t in Budget.TransactionModel,
+    from(t in Fund.TransactionModel,
       where:
         t.status in [:pending, :failed, :completed] and t.inserted_at < ^age_cutoff and
           t.inserted_at > ^lookback_cutoff
@@ -47,7 +47,7 @@ defmodule Systems.Budget.TransactionReconciliation do
   end
 
   defp reconcile_transaction(
-         %Budget.TransactionModel{transaction_id: nil, id: id, status: status},
+         %Fund.TransactionModel{transaction_id: nil, id: id, status: status},
          state
        ) do
     Logger.error(
@@ -57,12 +57,12 @@ defmodule Systems.Budget.TransactionReconciliation do
     record(state, :unresolvable, id, nil, status, nil, %{reason: "no provider uid"})
   end
 
-  defp reconcile_transaction(%Budget.TransactionModel{transaction_id: "free_" <> _}, state) do
+  defp reconcile_transaction(%Fund.TransactionModel{transaction_id: "free_" <> _}, state) do
     State.tally(state, :verified)
   end
 
   defp reconcile_transaction(
-         %Budget.TransactionModel{transaction_id: uid, id: id, status: status} = transaction,
+         %Fund.TransactionModel{transaction_id: uid, id: id, status: status} = transaction,
          state
        ) do
     case Payment.Public.reconcile_get_transaction(state, uid) do
@@ -87,7 +87,7 @@ defmodule Systems.Budget.TransactionReconciliation do
 
   defp apply_status(
          state,
-         %Budget.TransactionModel{status: :completed},
+         %Fund.TransactionModel{status: :completed},
          _provider_status,
          _raw_status
        ),
@@ -95,7 +95,7 @@ defmodule Systems.Budget.TransactionReconciliation do
 
   defp apply_status(
          state,
-         %Budget.TransactionModel{id: id, transaction_id: uid, status: status} = transaction,
+         %Fund.TransactionModel{id: id, transaction_id: uid, status: status} = transaction,
          provider_status,
          raw_status
        ) do
@@ -112,7 +112,7 @@ defmodule Systems.Budget.TransactionReconciliation do
   end
 
   defp resolve(%{transaction_id: uid}, :completed) do
-    case Budget.Public.complete_transaction(uid) do
+    case Fund.Public.complete_transaction(uid) do
       {:ok, _} -> {:ok, :resolved_completed}
       other -> {:error, other}
     end
@@ -121,7 +121,7 @@ defmodule Systems.Budget.TransactionReconciliation do
   end
 
   defp resolve(%{status: :pending, transaction_id: uid}, :failed) do
-    case Budget.Public.fail_transaction(uid) do
+    case Fund.Public.fail_transaction(uid) do
       {:ok, _} -> {:ok, :resolved_failed}
       {:error, :already_completed} -> {:ok, :verified}
       other -> {:error, other}
