+++ core/systems/payment/reconciliation_summary.ex
@@ -1,7 +1,7 @@
 defmodule Systems.Payment.ReconciliationSummary do
   @moduledoc """
   Tally of a reconciliation run, shared by the per-type reconcilers
-  (`Fund.Public.reconcile_pending_payouts/1`, `Budget.Public.reconcile_transactions/1`)
+  (`Fund.Public.reconcile_pending_payouts/1`, `Fund.Public.reconcile_transactions/1`)
   and aggregated by `Systems.Payment.ReconciliationWorker`.
 
   Outcomes:
