+++ core/test/systems/assignment/info_model_test.exs
@@ -49,7 +49,7 @@ defmodule Systems.Assignment.InfoModelTest do
     test "accepts a short description on auto_save" do
       # Short aims are allowed at auto-save time — we only guard against
       # overflow, not against low-effort descriptions. See
-      # Systems.Budget.PayInRequestModel for the not-empty guard at
+      # Systems.Fund.PayInRequestModel for the not-empty guard at
       # Confirm time.
       changeset = changeset(%{aim_of_study: "hi"})
