+++ core/test/systems/fund/factories.ex
@@ -73,8 +73,8 @@ defmodule Systems.Fund.Factories do
   end
 
   def insert_pay_in!(%Fund.Model{id: fund_id}, %Systems.Account.User{id: user_id}) do
-    %Budget.TransactionModel{}
-    |> Budget.TransactionModel.changeset(%{
+    %Fund.TransactionModel{}
+    |> Fund.TransactionModel.changeset(%{
       transaction_id: "tx_#{System.unique_integer([:positive])}",
       status: :completed,
       idempotence_key: Ecto.UUID.generate(),
